// fe/src/lib/api/generated/model/deliveryMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type DeliveryMode = typeof DeliveryMode[keyof typeof DeliveryMode];


export const DeliveryMode = {
  IN_PERSON: 'IN_PERSON',
  ONLINE: 'ONLINE',
} as const;