// fe/src/lib/api/generated/model/providerRoleCode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Response compatibility for historical provider-role assignments.
 */
export type ProviderRoleCode = typeof ProviderRoleCode[keyof typeof ProviderRoleCode];


export const ProviderRoleCode = {
  COUNSELOR: 'COUNSELOR',
  GUIDANCE_SERVICES_STAFF: 'GUIDANCE_SERVICES_STAFF',
} as const;