// fe/src/lib/api/generated/model/serviceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentPolicy } from './appointmentPolicy';
import type { DeliveryMode } from './deliveryMode';
import type { ProviderRoleCode } from './providerRoleCode';

export interface ServiceResponse {
  appointment_policy: AppointmentPolicy;
  cancellation_cutoff_minutes: number | null;
  code: string;
  created_at: string;
  default_duration_minutes: number | null;
  delivery_modes: DeliveryMode[];
  description: string;
  id: string;
  is_active: boolean;
  name: string;
  provider_roles: ProviderRoleCode[];
  requires_current_inventory: boolean;
  updated_at: string;
}