// fe/src/lib/api/generated/model/serviceUpdateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentPolicy } from './appointmentPolicy';
import type { ConfigurableProviderRoleCode } from './configurableProviderRoleCode';
import type { DeliveryMode } from './deliveryMode';

export interface ServiceUpdateRequest {
  appointment_policy?: AppointmentPolicy;
  cancellation_cutoff_minutes?: number | null;
  default_duration_minutes?: number | null;
  delivery_modes?: DeliveryMode[];
  description?: string;
  name?: string;
  provider_roles?: ConfigurableProviderRoleCode[];
  requires_current_inventory?: boolean;
}